rename testcase to group, to reduce name conflicts
